requirements_midnight_stretch.txt
actionlib==1.12.1
addict==2.4.0
angles==1.9.12
anyio==3.3.3
apturl==0.5.2
argon2-cffi==21.1.0
asn1crypto==0.24.0
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
base_local_planner==1.16.7
bleach==4.1.0
bondpy==1.8.5
boto==2.49.0
Brlapi==0.6.6
camera_calibration==1.15.0
camera_calibration_parsers==1.11.13
catkin==0.7.29
catkin-pkg==0.4.23
certifi==2018.1.18
cffi==1.15.0
chardet==3.0.4
click==6.7
cma==3.1.0
colorama==0.3.7
command-not-found==0.3
contextvars==2.4
controller_manager==0.18.4
controller_manager_msgs==0.18.4
cryptography==2.1.4
cupshelpers==1.0
cv_bridge==1.13.0
cycler==0.10.0
dataclasses==0.8
ddynamic_reconfigure_python==0.0.1
decorator==5.0.9
defer==1.0.6
defusedxml==0.7.1
deprecation==2.1.0
device==0.1
diagnostic_analysis==1.9.7
diagnostic_common_diagnostics==1.9.7
diagnostic_updater==1.9.7
distro==1.6.0
distro-info===0.18ubuntu0.18.04.1
docutils==0.17.1
drawnow==0.72.5
dynamic_reconfigure==1.6.3
dynamixel_sdk==3.7.51
entrypoints==0.3
evdev==1.4.0
freetype-py==2.2.0
future==0.18.2
gazebo_plugins==2.8.7
gazebo_ros==2.8.7
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.16
gitdb==4.0.7
GitPython==3.1.18
hello-robot-stretch-body==0.1.13
hello-robot-stretch-body-tools-py3==0.0.8
hello-robot-stretch-tool-share==0.1.7
httplib2==0.9.2
idna==3.3
image_geometry==1.13.0
imageio==2.9.0
immutables==0.16
importlib-metadata==4.8.1
inputs==0.5
interactive-markers==1.11.5
ipykernel==5.5.6
ipython==5.5.0
ipython_genutils==0.2.0
ipywidgets==7.6.5
Jinja2==3.0.2
joblib==1.1.0
joint_state_publisher==1.12.15
joint_state_publisher_gui==1.12.15
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyter-packaging==0.10.6
jupyter-server==1.11.1
jupyterlab==3.2.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
kdl_parser_py==1.13.1
keyboard==0.13.5
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
laser_geometry==1.6.7
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
llvmlite==0.36.0
louis==3.5.0
lxml==4.6.3
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==2.0.1
matplotlib==3.3.4
message_filters==1.14.11
mistune==0.8.4
moveit-core==1.0.8
moveit_ros_planning_interface==1.0.8
moveit_ros_visualization==1.0.8
mpmath==1.2.1
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.2
nose==1.3.7
notebook==6.4.4
numba==0.53.1
numpy==1.19.5
oauth==1.0.1
olefile==0.45.1
open3d==0.13.0
opencv-python-inference-engine==2021.9.10
openpifpaf==0.13.0
packaging==21.0
pandas==1.1.5
pandocfilters==1.5.0
pathlib==1.0.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==8.4.0
pixel-ring==0.1.0
playsound==1.2.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.0.0
psutil==5.8.0
ptyprocess==0.7.0
PyAudio==0.2.11
pycairo==1.16.2
pycollada==0.6
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pyglet==1.5.20
Pygments==2.10.0
PyGObject==3.26.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pynput==1.7.4
PyOpenGL==3.1.0
pyparsing==2.4.7
pyrealsense2==2.50.0.3812
pyrender==0.1.45
pyRFC3339==1.0
pyrsistent==0.18.0
pyserial==3.5
pysparkling==0.6.1
python-apt==1.6.5+ubuntu0.6
python-dateutil==2.8.2
python-debian==0.1.32
python-json-logger==2.0.2
python-xlib==0.31
python_qt_binding==0.4.4
pytz==2021.3
pyusb==1.2.1
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==6.0
pyzmq==22.3.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
qtconsole==5.1.1
QtPy==1.11.2
reportlab==3.4.0
requests==2.0.0
requests-unixsocket==0.1.5
resource_retriever==1.12.6
ros_numpy==0.0.3
ros_speech_recognition==2.1.24
rosapi==0.11.13
rosbag==1.14.11
rosboost-cfg==1.14.9
rosbridge_library==0.11.13
rosbridge_server==0.11.13
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.11
roslaunch==1.14.11
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.11
rosmake==1.14.9
rosmaster==1.14.11
rosmsg==1.14.11
rosnode==1.14.11
rosparam==1.14.11
rospkg==1.3.0
rospy==1.14.11
rosservice==1.14.11
rostest==1.14.11
rostopic==1.14.11
rosunit==1.14.9
roswtf==1.14.11
rpl==1.5.6
rplidar-roboticia==0.9.5
rqt-moveit==0.5.10
rqt-reconfigure==0.5.4
rqt-robot-monitor==0.5.13
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.9
rqt_dep==0.4.9
rqt_graph==0.4.11
rqt_gui==0.5.2
rqt_gui_py==0.5.2
rqt_image_view==0.4.16
rqt_launch==0.4.8
rqt_logger_level==0.4.8
rqt_msg==0.4.8
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.8
rqt_publisher==0.4.8
rqt_py_common==0.5.2
rqt_py_console==0.4.8
rqt_robot_dashboard==0.5.7
rqt_robot_steering==0.5.10
rqt_runtime_monitor==0.5.7
rqt_service_caller==0.4.8
rqt_shell==0.4.9
rqt_srv==0.4.8
rqt_tf_tree==0.6.0
rqt_top==0.4.8
rqt_topic==0.4.11
rqt_web==0.4.8
rviz==1.13.18
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
SecretStorage==2.3.1
Send2Trash==1.8.0
sensor-msgs==1.12.8
simplegeneric==0.8.1
simplejson==3.13.2
six==1.11.0
smach==2.0.1
smach_ros==2.0.1
smclib==1.8.5
smmap==4.0.0
snakeviz==2.1.0
sniffio==1.2.0
sound-play==0.3.11
SpeechRecognition==3.8.1
spidev==3.5
srdfdom==0.5.2
sympy==1.9
system-service==0.3
systemd-python==234
terminado==0.12.1
testpath==0.5.0
tf==1.12.1
tf2_geometry_msgs==0.6.5
tf2_kdl==0.6.5
tf2_py==0.6.5
tf2_ros==0.6.5
tf_conversions==1.12.1
threadpoolctl==3.0.0
tifffile==2020.9.3
tomlkit==0.7.2
topic_tools==1.14.11
torch==1.9.1
torchvision==0.10.1
tqdm==4.62.3
traitlets==4.3.2
trimesh==3.6.38
twilio==7.2.0
typing-extensions==3.10.0.2
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urdfdom-py==0.4.5
urdfpy==0.0.22
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==1.2.1
widgetsnbextension==3.5.1
xacro==1.13.12
xkit==0.0.0
zipp==3.6.0
zope.interface==4.3.2
